fix too big value ex fee (#37)

+++ modules/db-writer/src/main/scala/fi/spectrumlabs/db/writer/classes/Handle.scala
@@ -20,13 +20,12 @@ import fi.spectrumlabs.db.writer.services.Tokens
 import io.circe.parser._
 import io.circe.{Decoder, Encoder}
 import mouse.any._
-import org.scalatest.time.SpanSugar.convertIntToGrainOfTime
 import tofu.logging.{Logging, Logs}
 import tofu.syntax.foption.noneF
 import tofu.syntax.logging._
 import tofu.syntax.monadic._
 
-import scala.concurrent.duration.FiniteDuration
+import scala.concurrent.duration.{DurationInt, FiniteDuration}
 
 /** Keeps both ToSchema from A to B and Persist for B.
   * Contains evidence that A can be mapped into B and B can be persisted.
@@ -346,7 +345,7 @@ object Handle {
         )
         pool <- OptionT.liftF(tryGetPool(5, poolIn.txInRef, tx.txId))
         actualOutputOpt = pool.map(_.outputAmount(swap.base, swap.baseAmount.value))
-        fee             = actualOutputOpt.map(actualOutput => BigInt(actualOutput) * swap.exFeePerTokenNum / swap.exFeePerTokenDen)
+        fee             = actualOutputOpt.map(actualOutput => BigDecimal(actualOutput) * swap.exFeePerTokenNum / swap.exFeePerTokenDen)
 //        refundable      = swap.originalAdaAmount - fee
         aq = actualOutputOpt match {
           case Some(value) => if (swap.quote == Coin.Ada) value else actualQuote._2
